Version 1.0.5

+++ scripts/keyFunctions.js
@@ -1,44 +1,55 @@
-import { infoDisplay } from "./globalStore.js";
-import { retrieveData } from "./fetchFunctions.js";
+// UPDATET TO 1.0.5 //
+import { Elements } from "./globalStore.js";
 
 export function isKeyValid(key) {
-    return /^m=\d+&mk=[a-f0-9]{40}$/i.test(key);
+    // Check if the key matches the required pattern
+    return typeof key === 'string' && /^m=\d+&mk=[a-f0-9]{40}$/i.test(key);
 }
 
 export function trimKey(key) {
-    return key.split('=')[1].split('&')[0];
+    if (typeof key !== 'string') {
+        console.error('Expected a string for key trimming.');
+        return '';
+    }
+    
+    const parts = key.split('=');
+    return parts[1] ? parts[1].split('&')[0] : '';
 }
 
-export function badKeyDay(){
-    infoDisplay.innerHTML += "<h3 class='red'> Bad Key .. 😢</h3><button id='reload'> Reload </button>";
-
-    let reload = document.getElementById('reload');
+export function badKeyDay() {
+    Elements.infoDisplay.innerHTML += "<h3 class='red'> Bad Key .. 😢</h3><button id='reload'> Reload </button>";
     
-    reload.addEventListener('click', () =>{
-        localStorage.clear();
-        location.reload(true);
-    })
+    const reload = document.getElementById('reload');
+    
+    if (reload) {
+        reload.addEventListener('click', () => {
+            localStorage.clear();
+            location.reload(true);
+        });
+    }
 }
 
-export function checkSaveKeyInput(){
-    let _key = document.getElementById("settings-api-key").value
+export function checkSaveKeyInput() {
+    const apiKeyInput = document.getElementById("settings-api-key");
+    const saveButton = document.getElementById("settings-api-save");
+    const isValid = isKeyValid(apiKeyInput.value);
     
-    if(isKeyValid(_key)){
-        document.getElementById("settings-api-key").style.color = "green"
-        document.getElementById("settings-api-save").disabled = false;
-        document.getElementById("settings-api-save").style.backgroundColor = '#e73d3d';
-    }else{
-        document.getElementById("settings-api-key").style.color = "red"
-        document.getElementById("settings-api-save").disabled = true;
-        document.getElementById("settings-api-save").style.backgroundColor = '#5e5d5d';
-        
-    };
+    apiKeyInput.style.color = isValid ? "green" : "red";
+    saveButton.disabled = !isValid;
+    saveButton.style.backgroundColor = isValid ? '#e73d3d' : '#5e5d5d';
 }
 
 // Save new key and refresh data
 export function saveNewKeyandRefresh() {
-    const _accessKey = document.getElementById('settings-api-key').value;
-    localStorage.setItem('key', _accessKey);
-    window.location.reload();
+    const apiKeyInput = document.getElementById('settings-api-key');
+    const accessKey = apiKeyInput.value;
+    
+    if (isKeyValid(accessKey)) {
+        localStorage.setItem('key', accessKey);
+        window.location.reload();
+    } else {
+        console.error('Attempted to save an invalid key.');
+    }
 }
+
